Reorganize sign up feature specs
**Why**:
* Previously, "confirmation specs" where in their own dir
* There were also some confirmation specs in the sign up file
* This PR moves all of the sign-up related specs, including email
  confirmation, into the `spec/features/visitors` dir
* Also moved the `app_settings` feature specs into controller specs bc
  those are faster and the way we were testing app settings did not
  require a feature-level spec
* The goal of this PR is to make specs easier to find by having better
  organization
