+++ spec/features/visitors/navigation_spec.rb
@@ -1,6 +1,6 @@
 require 'rails_helper'
 
-feature 'Navigation links', devise: true do
+feature 'Navigation links' do
   scenario 'view navigation links' do
     visit root_path
     expect(page).to have_content t('links.sign_in')
