+++ spec/features/visitors/resend_email_confirmation_spec.rb
@@ -1,7 +1,7 @@
 require 'rails_helper'
 require 'email_spec'
 
-feature 'Confirmation Instructions', devise: true do
+feature 'Visit requests confirmation instructions again during sign up' do
   include(EmailSpec::Helpers)
   include(EmailSpec::Matchers)
