+++ spec/features/visitors/set_password_spec.rb
@@ -1,0 +1,99 @@
+require 'rails_helper'
+
+feature 'Visitor sets password during signup' do
+  scenario 'visitor is redirected back to password form when password is blank' do
+    create(:user, :unconfirmed)
+    confirm_last_user
+    fill_in 'password_form_password', with: ''
+    click_button t('forms.buttons.submit.default')
+
+    expect(page).to have_content t('errors.messages.blank')
+    expect(current_url).to eq confirm_url
+  end
+
+  context 'password field is blank when JS is on', js: true do
+    before do
+      create(:user, :unconfirmed)
+      confirm_last_user
+    end
+
+    it 'shows error message when password is blank' do
+      fill_in 'password_form_password', with: ''
+      click_button t('forms.buttons.submit.default')
+
+      expect(page).to have_content 'Please fill in this field.'
+    end
+  end
+
+  scenario 'password strength indicator hidden when JS is off' do
+    create(:user, :unconfirmed)
+    confirm_last_user
+
+    expect(page).to have_css('#pw-strength-cntnr.hide')
+  end
+
+  context 'password strength indicator when JS is on', js: true do
+    before do
+      create(:user, :unconfirmed)
+      confirm_last_user
+    end
+
+    it 'is visible on page (not have "hide" class)' do
+      expect(page).to_not have_css('#pw-strength-cntnr.hide')
+    end
+
+    it 'updates as password changes' do
+      expect(page).to have_content '...'
+
+      fill_in 'password_form_password', with: 'password'
+      expect(page).to have_content 'Very weak'
+
+      fill_in 'password_form_password', with: 'this is a great sentence'
+      expect(page).to have_content 'Great!'
+    end
+
+    it 'has dynamic password strength feedback' do
+      expect(page).to have_content '...'
+
+      fill_in 'password_form_password', with: 'password'
+      expect(page).to have_content 'This is a top-10 common password'
+    end
+  end
+
+  scenario 'password visibility toggle when JS is on', js: true do
+    create(:user, :unconfirmed)
+    confirm_last_user
+
+    expect(page).to have_css('input.password[type="password"]')
+
+    find('#pw-toggle-0', visible: false).trigger('click')
+
+    expect(page).to_not have_css('input.password[type="password"]')
+    expect(page).to have_css('input.password[type="text"]')
+  end
+
+  context 'password is invalid' do
+    scenario 'visitor is redirected back to password form' do
+      create(:user, :unconfirmed)
+      confirm_last_user
+      fill_in 'password_form_password', with: 'Q!2e'
+
+      click_button t('forms.buttons.submit.default')
+
+      expect(page).to have_content('characters')
+      expect(current_url).to eq confirm_url
+    end
+
+    scenario 'visitor gets password help message' do
+      allow(Figaro.env).to receive(:password_strength_enabled).and_return('true')
+
+      create(:user, :unconfirmed)
+      confirm_last_user
+      fill_in 'password_form_password', with: 'password'
+
+      click_button t('forms.buttons.submit.default')
+
+      expect(page).to have_content('not strong enough')
+    end
+  end
+end
